Un-nest StripHandledMemberDeclarationsRewriter into its own source file
Dedent the nested type to compilation-unit scope; type body is otherwise unchanged.
